:recycle: (entities, user.repository.ts) 프로젝트 파일구조 변경
프로젝트의 파일구조를 도메인형으로 변경, 변경에 따른 import 구문 변경
