+++ src/answer-photo/AnswerPhoto.entity.ts
@@ -1,17 +1,17 @@
 import { Column, Entity, JoinColumn, ManyToOne, PrimaryGeneratedColumn } from 'typeorm';
-import Question from './Question.entity';
+import Answer from '../answer/Answer.entity';
 
-@Entity('question_photo')
-export default class QuestionPhoto {
+@Entity('answer_photo')
+export default class AnswerPhoto {
   @PrimaryGeneratedColumn()
   idx: number;
 
   @Column({ name: 'image_path', nullable: true })
   imagePath: string | null;
 
-  @ManyToOne(() => Question, (question) => question.questionPhoto, {
+  @ManyToOne(() => Answer, (question) => question.answerPhoto, {
     onDelete: 'CASCADE'
   })
-  @JoinColumn({ name: 'question_idx' })
-  question: Question[];
+  @JoinColumn({ name: 'answer_idx' })
+  answer: Answer;
 }
