+++ src/user/User.entity.ts
@@ -9,8 +9,8 @@ import {
   PrimaryGeneratedColumn,
 } from 'typeorm';
 import * as bcrypt from 'bcrypt';
-import Question from './Question.entity';
-import Answer from './Answer.entity';
+import Question from '../question/Question.entity';
+import Answer from '../answer/Answer.entity';
 @Entity('user')
 export default class User extends BaseEntity {
   @PrimaryGeneratedColumn()
