+++ src/user/user.repository.ts
@@ -1,5 +1,5 @@
 import { EntityRepository, Repository } from 'typeorm';
-import User from '../entities/User.entity';
+import User from './User.entity';
 
 @EntityRepository(User)
 export class UserRepository extends Repository<User> {
